Separate `style.scss` into one for shared, one for comparison-grid, and one for features-grid.
